requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
google-auth==2.17.1
google-auth-oauthlib==1.0.0
openai==0.27.2
mysql-connector-python==8.0.32
gunicorn==20.1.0